+++ app/models/user.rb
@@ -1,5 +1,5 @@
 class User < ApplicationRecord
-  include GraphQL::Interface
+  include Interface
   # Include default devise modules. Others available are:
   # :timeoutable, :trackable and :omniauthable
   devise :database_authenticatable,
