+++ config/routes.rb
@@ -1,9 +1,12 @@
 Rails.application.routes.draw do
   devise_for :users, skip: :sessions
-  root 'home#index'
+
   if Rails.env.development?
     mount GraphiQL::Rails::Engine, at: "/graphiql", graphql_path: "/graphql"
   end
+
   post "/graphql", to: "graphql#execute"
+
+  root 'home#index'
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
